refactor(frontend): delete redundant MyPraisePageLoader, refactor PraisePageLoader to support my praise query, and MyPraiseTable to use
